GForge #19573
https://gforge.nci.nih.gov/tracker/index.php?func=detail&aid=19573&group_id=491&atid=1850

In the method 'getEntityStmtText' in the Class SQLIndexer.java,
it is building a SQL statement for the indexer to use. This SQL
statement joins a few tables, two of which being 'entity' and
'entityProperty' tables. Both of these tables have an 'isActive'
column, so during the join asking for the 'isActive' column was
ambiguous -- it depended on the order the underlying database
joined the tables. I removed 'isActive' from the 'entityProperty'
table to remove the ambiguity (it wasn't being used anyway).
Tests now pass in Postgres and MS Access.