Replaced the slightly older LexGrid200901.eap with the contents LexGrid200901_mod.eap, as the changes appear to have met general approval.
Added a note to entityVersion and codingSchemeVersion to make it doubly clear that they are not to be used
Changed the cardinality of revision.changeAgent from 0..1 to 0..*
Added a note to the changedEntry/versionable link to make it clear that it was derived.
changed a mislabeled role on the definitionEntry/entryReference link from "code" to entityReference"
Made pickText and entityCode mandatory in pickListEntry and changed the documentation to reflect the fact that this element no longer carries "placeholder" pick list entries.
Added documentation to "concept" and "instance" to clarify their actual intent and to (hopefully) avoid questions about why "concept" and "instance" appear in the model.
Added documentation to urnMap to make it clear that it doesn't align with the XML Schema.